refactor: fix some bugs in wallet service related to auctions

+++ internal/guild/entity.go
@@ -2,8 +2,8 @@ package guild
 
 type Guild struct {
 	ID         string
-	Gold       float64
-	Reserved   float64
-	DailyLimit float64 // max gold committable per day
-	DailySpent float64 // gold committed today (purchases + reservations)
+	Gold       int64
+	Reserved   int64
+	DailyLimit int64 // max gold committable per day
+	DailySpent int64 // gold committed today (purchases + reservations)
 }
